:recycle: Moved logging out to separate context
