+++ aksel.nav.no/website/components/layout/header/search/hooks/useSearch.ts
@@ -4,6 +4,7 @@ import { useMemo, useState } from "react";
 import useSWRImmutable from "swr/immutable";
 import { createSearchResult, formatResults, fuseSearch } from "../utils";
 import { debounce } from "@navikt/ds-react";
+import { logSearch } from "@/utils";
 
 export const useSearch = () => {
   const [fuseResults, setFuseResults] = useState<SearchResultsT>(null);
@@ -32,6 +33,12 @@ export const useSearch = () => {
         );
 
         setFuseResults(createSearchResult(formatedResults, rawResults));
+        logSearch({
+          type: "standard",
+          searchedFromUrl: window.location.pathname,
+          query: value,
+          filter: tags,
+        });
       }),
     [data]
   );
