+++ aksel.nav.no/website/components/layout/header/search/providers/index.ts
@@ -1,2 +1,3 @@
 export * from "./SearchProvider";
 export * from "./SearchResultProvider";
+export * from "./SearchLoggingProvider";
